Split the theme switcher into additional components for multiple rendering
