+++ docs/src/components/Navi.astro
@@ -1,7 +1,8 @@
 ---
 import Icon from "../components/Icon.vue";
 import { currentPage } from "../modules/currentPage";
-import ThemeSwitcher from "../components/ThemeSwitcher.vue";
+import ThemeSwitcherPopover from "../components/ThemeSwitcherPopover.vue";
+import ThemeSwitcherMenu from "../components/ThemeSwitcherMenu.vue";
 
 interface Props {
   classes?: string;
@@ -56,19 +57,22 @@ const cp = currentPage(new URL(Astro.request.url).pathname, {
   </li>
   {
     inline && (
-      <>
-        <li class="menu__sep" />
-        <li class={"menu__item margin-left-md"}>
-          <ThemeSwitcher client:only />
-        </li>
-      </>
+      <li class={"menu__item margin-left-md"}>
+        <ThemeSwitcherPopover client:only />
+      </li>
     )
   }
   {
     !inline && (
-      <li class="menu__item">
-        <p>Drawer theme switcher...</p>
-      </li>
+      <>
+        <li class="menu__sep" />
+        <li class="menu__item padding-x-md padding-y-sm font-weight-semi-bold">
+          Switch Theme
+        </li>
+        <li class="menu__item">
+          <ThemeSwitcherMenu client:only />
+        </li>
+      </>
     )
   }
 </ul>
